final_project/requirements.txt
streamlit
matplotlib
pandas
numpy
plotly
psycopg2-binary
